requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
mysql-connector-python>=8.1.0
sqlalchemy>=2.0.20
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.3.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic[email]>=2.3.0
bcrypt==3.2.0